Overhauled section management by switching to a tree model, adding additional possibilities for section management and implementation
The ultimate goal of the Section concept is still to allow different plugins to work nice together.
Sections are now organized as a tree, calling preparation and/or cleanup code of all sections on the tree path explored when player is moving between different sections.
Verification checks and actual preparation/cleanup logic are now better separated in the sections as the verification checks are called for all sections that need to be walked through before the actual preparation/cleanup logic is called. If implemented correctly by all sections, it ensures that no irreversible logic was called if one section refuse the player to move for an arbitrary reason.
While the old manager let all edge cases to be handled by calling code, the new manager handles most of them by itself. Most importantly, no explicit knowledge of hubs is required as now, when the manager is called without a section, it will move the player up the tree instead of trying to call the section's "preferred next section" (a concept which does not exists anymore) and failing if there is not, previously forcing the plugin to explicitly move the player to a hub.

Sections can now tune the manager's behavior with capabilities indicating which aspect they are aware of/can handle, allowing both the manager to apply default behavior when it is not the case, and the section to use custom logic, free of the manager's actions, when it is the case. This allows simple cases, notably everything regarding parties, to not be explicitly defined and replicated between each sections.